backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg[binary]>=3.2.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1
pyjwt>=2.9.0
passlib[bcrypt]>=1.7.4
pytest>=8.3.0
httpx>=0.27.0
ruff>=0.6.0
